+++ main.py
@@ -1,10 +1,10 @@
 from tkinter import *
 from tkinter import filedialog as f
-from tkinter import messagebox
-from io import open
-import os
 import platform
 import sys
+import os
+# package imports
+import src.menu.about as about
 
 TITLE = "Bloc de notas | Python"
 
@@ -54,29 +54,6 @@ def save_file():
                 url_file = ""
                 root.title("Guardado cancelado " + url_file + TITLE) 
 
-def display_license():
-    with open("LICENSE", "r") as file:
-        m = file.read()
-    messagebox.showinfo("License", m)
-
-def display_credits():
-    LINK = "https://github.com/janselroa/bloc-de-notas-python"
-    m = f"""
-Esta aplicación ha sido desarroyada por Jansel Roa en su repositorio de GitHub:
-
-{LINK}
-
-Puedes leer la lista completa de contribuidores de esta aplicación en el archivo contributors.md
-    """
-    messagebox.showinfo("Créditos", m)
-
-def display_contact():
-    m = """
-Si necesitas contactar en relacción con este programa, puedes hacerlo en nuestro repositorio de GitHub.
-
-Tambien puedes contactar a JanselRoa en twitter como @RoaJansel.
-"""
-    messagebox.showinfo("Contacto", m)
 
 
 # Menú
@@ -88,9 +65,9 @@ file_menu.add_command(label="Guardar archivo", command=save_file)
 file_menu.add_command(label="Salir", command=root.quit)
 bar.add_cascade(menu=file_menu, label="Archivo")
 more_menu = Menu(bar, tearoff=0)
-more_menu.add_command(label="Licencia", command=display_license)
-more_menu.add_command(label="Créditos", command=display_credits)
-more_menu.add_command(label="Contacto", command=display_contact)
+more_menu.add_command(label="Licencia", command=about.display_license)
+more_menu.add_command(label="Créditos", command=about.display_credits)
+more_menu.add_command(label="Contacto", command=about.display_contact)
 bar.add_cascade(menu=more_menu, label="Sobre")
 
 # caja de text, donde se escribe ._.XD
